+++ quickwit-cli/tests/cli.rs
@@ -30,7 +30,7 @@ use helpers::{TestEnv, TestStorageType};
 use predicates::prelude::*;
 use quickwit_cli::index::{create_index_cli, CreateIndexArgs};
 use quickwit_common::rand::append_random_suffix;
-use quickwit_metastore::{Metastore, MetastoreUriResolver};
+use quickwit_metastore::{quickwit_metastore_uri_resolver, Metastore};
 use serde_json::{Number, Value};
 use serial_test::serial;
 use tokio::time::{sleep, Duration};
@@ -366,9 +366,20 @@ async fn test_cmd_garbage_collect_no_grace() -> Result<()> {
         test_env.data_dir_path.as_path(),
     );
 
-    let metastore = MetastoreUriResolver::default()
+    let metastore = quickwit_metastore_uri_resolver()
         .resolve(&test_env.metastore_uri)
         .await?;
+
+    let refresh_metastore = |metastore| {
+        // In this test we rely on the file backed metastore and write on
+        // a different process. The file backed metastore caches results.
+        // Therefore we need to force reading the disk.
+        //
+        // We do that by dropping and recreating our metastore.
+        drop(metastore);
+        quickwit_metastore_uri_resolver().resolve(&test_env.metastore_uri)
+    };
+
     let splits = metastore.list_all_splits(&test_env.index_id).await?;
     assert_eq!(splits.len(), 1);
     make_command(
@@ -388,9 +399,11 @@ async fn test_cmd_garbage_collect_no_grace() -> Result<()> {
     assert_eq!(index_path.exists(), true);
 
     let split_ids = [splits[0].split_id()];
+    let metastore = refresh_metastore(metastore).await?;
     metastore
         .mark_splits_for_deletion(&test_env.index_id, &split_ids)
         .await?;
+
     make_command(
         format!(
             "index gc --index-id {} --metastore-uri {} --dry-run --grace-period 10m",
@@ -431,9 +444,7 @@ async fn test_cmd_garbage_collect_no_grace() -> Result<()> {
         assert_eq!(split_filepath.exists(), false);
     }
 
-    let metastore = MetastoreUriResolver::default()
-        .resolve(&test_env.metastore_uri)
-        .await?;
+    let metastore = refresh_metastore(metastore).await?;
     assert_eq!(
         metastore.list_all_splits(&test_env.index_id).await?.len(),
         0
