+++ quickwit-common/src/lib.rs
@@ -19,7 +19,10 @@
 
 mod checklist;
 mod coolid;
+pub mod fs;
 pub mod metrics;
+pub mod net;
+pub mod rand;
 pub mod uri;
 
 use std::fmt::Debug;
@@ -96,104 +99,5 @@ mod tests {
         assert_eq!(super::get_from_env(TEST_KEY, 10), 15);
         std::env::set_var(TEST_KEY, "1invalidnumber");
         assert_eq!(super::get_from_env(TEST_KEY, 10), 10);
-    }
-}
-
-pub mod fs {
-    use std::path::Path;
-
-    use tokio;
-
-    /// Deletes the contents of a directory.
-    pub async fn empty_dir<P: AsRef<Path>>(path: P) -> anyhow::Result<()> {
-        let mut entries = tokio::fs::read_dir(path).await?;
-        while let Some(entry) = entries.next_entry().await? {
-            if entry.file_type().await?.is_dir() {
-                tokio::fs::remove_dir_all(entry.path()).await?
-            } else {
-                tokio::fs::remove_file(entry.path()).await?;
-            }
-        }
-        Ok(())
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use tempfile;
-
-        use super::*;
-
-        #[tokio::test]
-        async fn test_empty_dir() -> anyhow::Result<()> {
-            let tempdir = tempfile::tempdir()?;
-
-            let file_path = tempdir.path().join("file");
-            tokio::fs::File::create(file_path).await?;
-
-            let subdir = tempdir.path().join("subdir");
-            tokio::fs::create_dir(&subdir).await?;
-
-            let subfile_path = subdir.join("subfile");
-            tokio::fs::File::create(subfile_path).await?;
-
-            empty_dir(tempdir.path()).await?;
-            assert!(tokio::fs::read_dir(tempdir.path())
-                .await?
-                .next_entry()
-                .await?
-                .is_none());
-            Ok(())
-        }
-    }
-}
-
-pub mod net {
-    use std::net::{IpAddr, Ipv4Addr, SocketAddr, TcpListener, ToSocketAddrs};
-
-    /// Finds a random available port.
-    pub fn find_available_port() -> anyhow::Result<u16> {
-        let socket = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1)), 0);
-        let listener = TcpListener::bind(socket)?;
-        let port = listener.local_addr()?.port();
-        Ok(port)
-    }
-
-    /// Converts this string to a resolved `SocketAddr`.
-    pub fn socket_addr_from_str<S: AsRef<str>>(addr: S) -> anyhow::Result<SocketAddr> {
-        addr.as_ref()
-            .to_socket_addrs()?
-            .next()
-            .ok_or_else(|| anyhow::anyhow!("Failed to resolve address `{}`.", addr.as_ref()))
-    }
-}
-
-pub mod rand {
-    use rand::distributions::Alphanumeric;
-    use rand::Rng;
-
-    /// Appends a random suffix composed of a hyphen and five random alphanumeric characters.
-    pub fn append_random_suffix(string: &str) -> String {
-        let rng = rand::thread_rng();
-        let slug: String = rng
-            .sample_iter(&Alphanumeric)
-            .take(5)
-            .map(char::from)
-            .collect();
-        format!("{}-{}", string, slug)
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use super::append_random_suffix;
-
-        #[test]
-        fn test_append_random_suffix() -> anyhow::Result<()> {
-            let randomized = append_random_suffix("");
-            let mut chars = randomized.chars();
-            assert_eq!(chars.next(), Some('-'));
-            assert_eq!(chars.clone().count(), 5);
-            assert!(chars.all(|ch| ch.is_ascii_alphanumeric()));
-            Ok(())
-        }
     }
 }
