+++ quickwit-indexing/src/source/kafka_source.rs
@@ -44,7 +44,7 @@ use tokio::task::spawn_blocking;
 use tracing::{debug, info, warn};
 
 use crate::models::RawDocBatch;
-use crate::source::{IndexerMessage, Source, TypedSourceFactory};
+use crate::source::{IndexerMessage, Source, SourceContext, TypedSourceFactory};
 
 /// We try to emit chewable batches for the indexer.
 /// One batch = one message to the indexer actor.
